requirements.txt
aiogram==2.13
motor==2.3.0
aiohttp==3.7.4.post0
python-dotenv==0.18.0
dnspython==2.0.0
eth_account
bip39
